refactor(identity): rename UserRole property to Role
