+++ TeachBoard.Backend/TeachBoard.IdentityService/TeachBoard.IdentityService.Domain/Entities/User.cs
@@ -9,7 +9,7 @@ public class User : BaseEntity
     public string UserName { get; set; } = string.Empty;
     public string PasswordHash { get; set; } = string.Empty;
 
-    public Role Role { get; set; } = Role.Unspecified;
+    public UserRole Role { get; set; } = UserRole.Unspecified;
 
     public string FirstName { get; set; } = string.Empty;
     public string LastName { get; set; } = string.Empty;
